Don't clone job definitions.
`buffer` and `string` jobs can specify their lengths using the name of
another, previously parsed value. This is handled by copying the job
definition and replacing the job's length property.

Instead of copying the job, it's much better to just store the length
in a variable instead. This slightly improves the performance of these
job types and reduces the number of objects being created, thus improving
GC times.